requirements.txt
django
djangorestframework
python-dotenv
cryptography
django-ratelimit
celery
django-storages
boto3
requests
